refactor(backend): refactor directory structures
