+++ backend/src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
-import { NotesModule } from "./controllers/notes/notes.module";
-import { TagsNotesModule } from "./controllers/tags/notes/tags-notes.module";
-import { TagsModule } from "./controllers/tags/tags.module";
+import { NotesModule } from "./web/controllers/notes/notes.module";
+import { TagsNotesModule } from "./web/controllers/tags/notes/tags-notes.module";
+import { TagsModule } from "./web/controllers/tags/tags.module";
 
 @Module({
   imports: [NotesModule, TagsModule, TagsNotesModule],
