--- backend/src/controllers/notes/notes.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { NotesController } from "./notes.controller";
-import { NotesService } from "../../domain/notes/notes.service";
-
-describe("NotesController", () => {
-  let controller: NotesController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [NotesController],
-      providers: [NotesService],
-    }).compile();
-
-    controller = module.get<NotesController>(NotesController);
-  });
-
-  it("should be defined", () => {
-    expect(controller).toBeDefined();
-  });
-});
