--- backend/src/controllers/tags/notes/dto/cursor.input.ts
@@ -1,10 +1,0 @@
-import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";
-
-export class Cursor {
-  @ApiProperty({
-    type: "string",
-    format: "date-time",
-  })
-  @ApiPropertyOptional()
-  cursor: string | undefined;
-}
