+++ backend/src/web/controllers/notes/notes.controller.ts
@@ -7,10 +7,10 @@ import {
   Query,
 } from "@nestjs/common";
 import { ApiQuery, ApiTags } from "@nestjs/swagger";
-import { Cursor } from "./dto/cursor.input";
-import { Limit } from "./dto/limit.input";
-import { Order } from "./dto/order.input";
-import { NotesService } from "../../domain/notes/notes.service";
+import { Cursor } from "../../requests/notes/cursor.input";
+import { Limit } from "../../requests/notes/limit.input";
+import { Order } from "../../requests/notes/order.input";
+import { NotesService } from "../../../domain/notes/notes.service";
 
 @ApiTags("notes")
 @Controller("notes")
