+++ backend/src/web/controllers/tags/notes/tags-notes.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
-import { TagsService } from "../../../domain/tags/tags.service";
-import { NotesService } from "../../../domain/notes/notes.service";
+import { TagsService } from "../../../../domain/tags/tags.service";
+import { NotesService } from "../../../../domain/notes/notes.service";
 import { TagsNotesController } from "./tags-notes.controller";
 
 @Module({
