+++ backend/src/web/controllers/tags/tags.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { TagsController } from "./tags.controller";
-import { TagsService } from "../../domain/tags/tags.service";
+import { TagsService } from "../../../domain/tags/tags.service";
 
 describe("TagsController", () => {
   let controller: TagsController;
